reorg: merge files, use logging, adding config
