--- scripts/post_build/post_build.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-echo "   post-build steps: shell script"
