--- scripts/pre_build/pre_build.go
@@ -1,7 +1,0 @@
-package main
-
-import "fmt"
-
-func main() {
-	fmt.Printf("   pre-build steps : go main module\n")
-}
